squash code into one file
